+++ app/Modules/Reports/Controllers/ItemSalesReportController.php
@@ -14,7 +14,7 @@ namespace BT\Modules\Reports\Controllers;
 use BT\Http\Controllers\Controller;
 use BT\Modules\Reports\Reports\ItemSalesReport;
 use BT\Modules\Reports\Requests\DateRangeRequest;
-use BT\Support\PDF\PDFFactory;
+use BT\Support\domPDF;
 
 class ItemSalesReportController extends Controller
 {
@@ -49,7 +49,7 @@ class ItemSalesReportController extends Controller
 
     public function pdf()
     {
-        $pdf = PDFFactory::create();
+        $pdf = new domPDF();
 
         $results = $this->report->getResults(
             request('from_date'),